requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiotinydb==2.0.0
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
attrs==24.3.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
chatoperastore==1.2.0
click==8.1.7
cnsyn==1.2.0
cntext==1.9.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.2.0
Deprecated==1.2.15
dill==0.3.8
distro==1.9.0
evaluate==0.4.3
fastapi==0.115.5
ffmpy==0.5.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
gensim==4.3.3
gradio==5.15.0
gradio_client==1.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.28.1
idna==3.10
jieba==0.42.1
Jinja2==3.1.5
jiter==0.7.1
joblib==1.4.2
json_repair==0.32.0
kiwisolver==1.4.8
llvmlite==0.43.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
mittens==0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.55.1
orjson==3.10.15
pandas==2.2.3
pdfdocx==1.7
pillow==11.1.0
propcache==0.2.1
pyarrow==18.1.0
pydantic==2.10.1
pydantic_core==2.27.1
pydub==0.25.1
PyMuPDF==1.25.1
pyparsing==3.2.0
python-docx==1.1.2
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
rdflib==7.1.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rouge_score==0.1.2
ruff==0.9.4
safehttpx==0.1.6
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==3.3.1
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
synonyms==3.23.5
textdistance==4.6.3
threadpoolctl==3.5.0
tiktoken==0.8.0
tinydb==4.8.2
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
triton==3.1.0
typer==0.15.1
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==14.2
Whoosh==2.7.4
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.3
